implicit loss of integer bits made explicit
git-svn-id: http://wnd-charm.googlecode.com/svn/pychrm/branches/shared_ImageMatrix@624 d860a611-dd6a-fbde-3e4e-7fa7191099b5

+++ src/statistics/CombFirst4Moments.cpp
@@ -71,7 +71,7 @@ int matr4moments_to_hist(double matr4moments[4][N_COMB_SAMPLES], double *vec, in
 		for (bin_index = 0; bin_index < nbins; bin_index++)
 			vec[vec_index++] = bins[bin_index];
 	}
-	return(vec_index);
+	return((int)vec_index);
 }
 
 //---------------------------------------------------------------------------
